+++ workers/javascript/packages/server/src/language-step-executor/native-plugin/bootstrap.test.ts
@@ -20,7 +20,7 @@ describe('bootstrap', () => {
       };
       context.kvStore = mockStore as unknown as KVStore;
 
-      const code = `var path = require('path');
+      const code = `var path = await import('path');
         var _ = require('lodash');
 
         console.log("Starting script...");
@@ -69,7 +69,7 @@ describe('bootstrap', () => {
       };
       context.kvStore = mockStore as unknown as KVStore;
 
-      const code = `var path = require('path');
+      const code = `var path = await import('path');
         var _ = require('lodash');
 
         globalResult = path.join('/tmp', 'test_data', name);
@@ -199,7 +199,7 @@ describe('bootstrap', () => {
       const filePaths: Record<string, string> = {};
       const inheritedEnv: Array<string> = [];
 
-      const expectedErr = `Error on line 79:
+      const expectedErr = `Error on line 138:
 Error: variables not defined`;
 
       const result: ExecutionOutput = await executeCode({ context, code, filePaths, inheritedEnv });
@@ -307,7 +307,7 @@ ReferenceError: userName is not defined`;
       const filePaths: Record<string, string> = {};
       const inheritedEnv: Array<string> = [];
 
-      const expectedErr = `Error on line -36:
+      const expectedErr = `Error on line -20:
 Error: Cannot find module 'child_process'`;
 
       let code = `var childProc = require('child_process');
@@ -319,7 +319,7 @@ Error: Cannot find module 'child_process'`;
       expect(result.output).toEqual({});
       expect(result.error).toEqual(expectedErr);
 
-      const expectedErrNode = `Error on line -36:
+      const expectedErrNode = `Error on line -20:
 Error: Cannot find module 'node:child_process'`;
 
       code = `var childProc = require('node:child_process');
@@ -331,18 +331,120 @@ Error: Cannot find module 'node:child_process'`;
       expect(resultNode.output).toEqual({});
       expect(resultNode.error).toEqual(expectedErrNode);
     });
+
+    it('dynamic imports of child_process throw an error', async () => {
+      const mockStore = new MockKVStore();
+      await mockStore.write('name', 'TestFunc');
+      const context: ExecutionContext = new ExecutionContext();
+      context.variables = {
+        name: {
+          key: 'name',
+          type: VariableType.Native,
+          mode: 'read'
+        }
+      };
+      context.kvStore = mockStore as unknown as KVStore;
+      const filePaths: Record<string, string> = {};
+      const inheritedEnv: Array<string> = [];
+
+      const expectedErr = `Error on line -34:
+Error: Cannot find module 'child_process'`;
+
+      let code = `var childProc = await import('child_process');
+        console.log(childProc.execSync('echo $HOME').toString());
+            `;
+      const result: ExecutionOutput = await executeCode({ context, code, filePaths, inheritedEnv });
+
+      expect(result).toBeDefined();
+      expect(result.output).toEqual({});
+      expect(result.error).toEqual(expectedErr);
+
+      const expectedErrNode = `Error on line -34:
+Error: Cannot find module 'node:child_process'`;
+
+      code = `var childProc = await import('node:child_process');
+        console.log(childProc.execSync('env').toString());
+            `;
+      const resultNode: ExecutionOutput = await executeCode({ context, code, filePaths, inheritedEnv });
+
+      expect(resultNode).toBeDefined();
+      expect(resultNode.output).toEqual({});
+      expect(resultNode.error).toEqual(expectedErrNode);
+    });
   });
 
-  describe('Imports of process should return santized process object', () => {
-    const testCases: Array<[string, string[] | null | undefined, Record<string, string>, string[]]> = [
-      ['populated', ['VAR_1', 'VAR_2', 'VAR_3'], { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, ['VAR_1', 'VAR_3']],
-      ['empty', [], { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, []],
-      ['null', null, { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, []],
-      ['undefined', undefined, { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, []]
-    ];
+  describe('indirect imports should throw errors', () => {
+    const moduleName = "child_process";
 
-    for (const [testName, inheritedEnv, testEnv, expectedOutput] of testCases) {
-      test(`Process env for ${testName} inherited env`, async () => {
+    const testCases: Array<[string, string, RegExp]> = [
+      [
+        'require through module',
+        `module.require('${moduleName}')`,
+        /^Error on line -?\d+:\s+Error: Cannot find module '\w+'$/,
+      ],
+      [
+        'require through module.parent',
+        `module.parent.require('${moduleName}')`,
+        /^Error on line -?\d+:\s+Error: Cannot find module '\w+'$/,
+      ],
+      [
+        'dynamic import',
+        `await import('${moduleName}')`,
+        /^Error on line -?\d+:\s+Error: Cannot find module '\w+'$/,
+      ],
+      [
+        'dynamic import through eval',
+        `await eval("(async () => { return await import('${moduleName}') })()")`,
+        /^Error on line \d+:\s+ReferenceError: eval is not defined$/,
+      ],
+      [
+        'dynamic import through indirect eval',
+        `await ('', eval)("(async () => { return await import('${moduleName}') })()")`,
+        /^Error on line \d+:\s+ReferenceError: eval is not defined$/,
+      ],
+      [
+        'require through Function.constructor',
+        `Function.constructor("return require('${moduleName}')")()`,
+        /^Error on line \d+:\s+ReferenceError: Function is not defined$/,
+      ],
+      [
+        'dynamic import through Function.constructor',
+        `await Function.constructor("return import('${moduleName}')")()`,
+        /^Error on line \d+:\s+ReferenceError: Function is not defined$/,
+      ],
+      [
+        'require through anonymous function constructor',
+        `(() => {}).constructor("return require('${moduleName}')")()`,
+        /^Error on line \d+:\s+TypeError: \(?[\w\s]+\)?\.constructor is not a function$/,
+      ],
+      [
+        'dynamic import through anonymous function constructor',
+        `await (() => {}).constructor("return import('${moduleName}')")()`,
+        /^Error on line \d+:\s+TypeError: \(?[\w\s]+\)?\.constructor is not a function$/,
+      ],
+      [
+        'require through AsyncFunction.constructor',
+        `await AsyncFunction.constructor("return require('${moduleName}')")()`,
+        /^Error on line \d+:\s+ReferenceError: AsyncFunction is not defined$/,
+      ],
+      [
+        'dynamic import through AsyncFunction.constructor',
+        `await AsyncFunction.constructor("return import('${moduleName}')")()`,
+        /^Error on line \d+:\s+ReferenceError: AsyncFunction is not defined$/,
+      ],
+      [
+        'require through anonymous async function constructor',
+        `await (async () => {}).constructor("return require('${moduleName}')")()`,
+        /^Error on line \d+:\s+TypeError: \(?[\w\s]+\)?\.constructor is not a function$/,
+      ],
+      [
+        'dynamic import through anonymous async function constructor',
+        `await (async () => {}).constructor("return import('${moduleName}')")()`,
+        /^Error on line \d+:\s+TypeError: \(?[\w\s]+\)?\.constructor is not a function$/,
+      ],
+    ];
+    for (const [name, importExpr, expectedErr] of testCases) {
+      it(`should throw error when importing ${name}`, async () => {
         const mockStore = new MockKVStore();
         await mockStore.write('name', 'TestFunc');
         const context: ExecutionContext = new ExecutionContext();
@@ -355,31 +457,73 @@ Error: Cannot find module 'node:child_process'`;
         };
         context.kvStore = mockStore as unknown as KVStore;
         const filePaths: Record<string, string> = {};
+        const inheritedEnv: Array<string> = [];
 
-        const codeProcess = `var nodeProcess = require('process');
-          return Object.keys(nodeProcess.env);
-              `;
+        const code = `const childProc = ${importExpr};
+        console.log(childProc.execSync('echo $HOME').toString());
+            `;
+        const result: ExecutionOutput = await executeCode({ context, code, filePaths, inheritedEnv });
 
-        const existingEnv = process.env;
-        process.env = testEnv;
-
-        const resultProcess: ExecutionOutput = await executeCode({ context, code: codeProcess, filePaths, inheritedEnv });
-
-        expect(resultProcess).toBeDefined();
-        expect(resultProcess.error).not.toBeDefined();
-        expect(resultProcess.output).toEqual(expectedOutput);
-
-        const codeNodeProcess = `var nodeProcess = require('node:process');
-          return Object.keys(nodeProcess.env);
-              `;
-        const resultNodeProcess: ExecutionOutput = await executeCode({ context, code: codeNodeProcess, filePaths, inheritedEnv });
-
-        process.env = existingEnv;
-
-        expect(resultNodeProcess).toBeDefined();
-        expect(resultNodeProcess.output).toEqual(expectedOutput);
-        expect(resultNodeProcess.error).not.toBeDefined();
+        expect(result).toBeDefined();
+        expect(result.output).toEqual({});
+        expect(result.error).toMatch(expectedErr);
       });
+    }
+  });
+
+  describe('Imports of process should return santized process object', () => {
+    const testCases: Array<[string, string[] | null | undefined, Record<string, string>, string[]]> = [
+      ['populated', ['VAR_1', 'VAR_2', 'VAR_3'], { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, ['VAR_1', 'VAR_3']],
+      ['empty', [], { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, []],
+      ['null', null, { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, []],
+      ['undefined', undefined, { VAR_1: 'val1', VAR_3: 'val3', OTHER_VAR: 'otherValue' }, []]
+    ];
+
+    const importTypes: Array<[string, string]> = [
+      ['require', 'require'],
+      ['import', 'await import']
+    ];
+    for (const [importType, importKeyword] of importTypes) {
+      for (const [testName, inheritedEnv, testEnv, expectedOutput] of testCases) {
+        test(`${importType} ${testName} process env for inherited env`, async () => {
+          const mockStore = new MockKVStore();
+          await mockStore.write('name', 'TestFunc');
+          const context: ExecutionContext = new ExecutionContext();
+          context.variables = {
+            name: {
+              key: 'name',
+              type: VariableType.Native,
+              mode: 'read'
+            }
+          };
+          context.kvStore = mockStore as unknown as KVStore;
+          const filePaths: Record<string, string> = {};
+
+          const codeProcess = `var nodeProcess = ${importKeyword}('process');
+            return Object.keys(nodeProcess.env);
+                `;
+
+          const existingEnv = process.env;
+          process.env = testEnv;
+
+          const resultProcess: ExecutionOutput = await executeCode({ context, code: codeProcess, filePaths, inheritedEnv });
+
+          expect(resultProcess).toBeDefined();
+          expect(resultProcess.error).not.toBeDefined();
+          expect(resultProcess.output).toEqual(expectedOutput);
+
+          const codeNodeProcess = `var nodeProcess = ${importKeyword}('node:process');
+            return Object.keys(nodeProcess.env);
+                `;
+          const resultNodeProcess: ExecutionOutput = await executeCode({ context, code: codeNodeProcess, filePaths, inheritedEnv });
+
+          process.env = existingEnv;
+
+          expect(resultNodeProcess).toBeDefined();
+          expect(resultNodeProcess.output).toEqual(expectedOutput);
+          expect(resultNodeProcess.error).not.toBeDefined();
+        });
+      }
     }
   });
 
@@ -429,7 +573,7 @@ Error: Cannot find module 'node:child_process'`;
     const testCases: Array<[string, unknown[], unknown]> = [
       ['channel.ref', [], "Error on line 1:\nTypeError: Cannot read properties of undefined (reading 'ref')"],
       ['channel.unref', [], "Error on line 1:\nTypeError: Cannot read properties of undefined (reading 'unref')"],
-      ['chdir', [`'/root'`], 'Error on line 103:\nError: Unable to change directory in a sandboxed environment']
+      ['chdir', [`'/root'`], 'Error on line 163:\nError: Unable to change directory in a sandboxed environment']
     ];
 
     for (const [restrictedMethod, args, expectedErr] of testCases) {
@@ -474,6 +618,7 @@ Error: Cannot find module 'node:child_process'`;
           AWS_REGION: process.env.AWS_REGION
         }
       ],
+      ['mainModule', undefined],
       ['ppid', process.pid],
       ['send', undefined]
     ];